Utilizar IP pelo .env e implementar função para checar se o usuário já está logado nas telas iniciais
