+++ frontend/services/index.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const api = axios.create({
-  baseURL: 'http://localhost:3000/'
+  baseURL: `${process.env.EXPO_PUBLIC_API_URL}/`
 });
 
 export default api;
